Moved file system functions into src/backen/common/util.cpp
